+++ num_extract/data/example_of_table.png
